# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UPSC Prep Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions with optional filters
      parameters:
        - name: subject
          in: query
          schema: { type: string }
        - name: topic
          in: query
          schema: { type: string }
        - name: difficulty
          in: query
          schema: { type: string, enum: [easy, medium, hard] }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionListResponse"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a single MCQ question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/bulk:
    post:
      operationId: bulkCreateQuestions
      tags: [questions]
      summary: Bulk create MCQ questions from parsed CSV/JSON data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkQuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkCreateResult"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a question by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── TESTS ───────────────────────────────────────────────────────────────────
  /tests:
    get:
      operationId: listTests
      tags: [tests]
      summary: List all test series
      parameters:
        - name: subject
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [draft, published, archived] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Test"
    post:
      operationId: createTest
      tags: [tests]
      summary: Create a test series
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
  /tests/{id}:
    get:
      operationId: getTest
      tags: [tests]
      summary: Get a test with its questions
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestWithQuestions"
        "404":
          description: Not found
    patch:
      operationId: updateTest
      tags: [tests]
      summary: Update a test
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
        "404":
          description: Not found
    delete:
      operationId: deleteTest
      tags: [tests]
      summary: Delete a test
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /tests/{id}/questions:
    post:
      operationId: addQuestionsToTest
      tags: [tests]
      summary: Add questions to a test series
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestQuestionsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"

  # ─── ATTEMPTS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Question:
      type: object
      required: [id, text, optionA, optionB, optionC, optionD, correctOption, subject, difficulty]
      properties:
        id:            { type: integer }
        text:          { type: string }
        optionA:       { type: string }
        optionB:       { type: string }
        optionC:       { type: string }
        optionD:       { type: string }
        correctOption: { type: string, enum: [A, B, C, D] }
        explanation:   { type: ["string", "null"] }
        subject:       { type: string }
        topic:         { type: ["string", "null"] }
        difficulty:    { type: string, enum: [easy, medium, hard] }
        year:          { type: ["integer", "null"] }
        createdAt:     { type: string, format: date-time }
    QuestionInput:
      type: object
      required: [text, optionA, optionB, optionC, optionD, correctOption, subject, difficulty]
      properties:
        text:          { type: string, minLength: 1 }
        optionA:       { type: string, minLength: 1 }
        optionB:       { type: string, minLength: 1 }
        optionC:       { type: string, minLength: 1 }
        optionD:       { type: string, minLength: 1 }
        correctOption: { type: string, enum: [A, B, C, D] }
        explanation:   { type: string }
        subject:       { type: string, minLength: 1 }
        topic:         { type: string }
        difficulty:    { type: string, enum: [easy, medium, hard] }
        year:          { type: integer }
    QuestionUpdate:
      type: object
      properties:
        text:          { type: string, minLength: 1 }
        optionA:       { type: string, minLength: 1 }
        optionB:       { type: string, minLength: 1 }
        optionC:       { type: string, minLength: 1 }
        optionD:       { type: string, minLength: 1 }
        correctOption: { type: string, enum: [A, B, C, D] }
        explanation:   { type: string }
        subject:       { type: string }
        topic:         { type: string }
        difficulty:    { type: string, enum: [easy, medium, hard] }
        year:          { type: integer }
    QuestionListResponse:
      type: object
      required: [questions, total]
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total: { type: integer }
    BulkQuestionInput:
      type: object
      required: [questions]
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionInput"
    BulkCreateResult:
      type: object
      required: [created, failed]
      properties:
        created: { type: integer }
        failed:  { type: integer }
        errors:
          type: array
          items: { type: string }

    # ── Test ──────────────────────────────────────────────────────────────────
    Test:
      type: object
      required: [id, title, subject, duration, totalMarks, status, questionCount]
      properties:
        id:            { type: integer }
        title:         { type: string }
        description:   { type: ["string", "null"] }
        subject:       { type: string }
        duration:      { type: integer, description: "Duration in minutes" }
        totalMarks:    { type: integer }
        negativeMarks: { type: number }
        status:        { type: string, enum: [draft, published, archived] }
        questionCount: { type: integer }
        attemptCount:  { type: integer }
        scheduledFor:  { type: ["string", "null"], format: date }
        createdAt:     { type: string, format: date-time }
    TestInput:
      type: object
      required: [title, subject, duration, totalMarks]
      properties:
        title:         { type: string, minLength: 1 }
        description:   { type: string }
        subject:       { type: string, minLength: 1 }
        duration:      { type: integer, minimum: 1 }
        totalMarks:    { type: integer, minimum: 1 }
        negativeMarks: { type: number, minimum: 0 }
        status:        { type: string, enum: [draft, published, archived] }
        scheduledFor:  { type: string, format: date }
    TestUpdate:
      type: object
      properties:
        title:         { type: string, minLength: 1 }
        description:   { type: string }
        subject:       { type: string }
        duration:      { type: integer, minimum: 1 }
        totalMarks:    { type: integer, minimum: 1 }
        negativeMarks: { type: number, minimum: 0 }
        status:        { type: string, enum: [draft, published, archived] }
        scheduledFor:  { type: string, format: date }
    TestWithQuestions:
      type: object
      required: [id, title, subject, duration, totalMarks, status, questionCount, attemptCount, questions]
      properties:
        id:            { type: integer }
        title:         { type: string }
        description:   { type: ["string", "null"] }
        subject:       { type: string }
        duration:      { type: integer }
        totalMarks:    { type: integer }
        negativeMarks: { type: number }
        status:        { type: string, enum: [draft, published, archived] }
        questionCount: { type: integer }
        attemptCount:  { type: integer }
        scheduledFor:  { type: ["string", "null"], format: date }
        createdAt:     { type: string, format: date-time }
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
    TestQuestionsInput:
      type: object
      required: [questionIds]
      properties:
        questionIds:
          type: array
          items: { type: integer }

    # ── Attempt ───────────────────────────────────────────────────────────────
...