# lib/api-zod/src/generated/types/questionListResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * UPSC Prep Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from './question';

export interface QuestionListResponse {
  questions: Question[];
  total: number;
}