# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * UPSC Prep Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionUpdateCorrectOption } from './questionUpdateCorrectOption';
import type { QuestionUpdateDifficulty } from './questionUpdateDifficulty';

export interface QuestionUpdate {
  /** @minLength 1 */
  text?: string;
  /** @minLength 1 */
  optionA?: string;
  /** @minLength 1 */
  optionB?: string;
  /** @minLength 1 */
  optionC?: string;
  /** @minLength 1 */
  optionD?: string;
  correctOption?: QuestionUpdateCorrectOption;
  explanation?: string;
  subject?: string;
  topic?: string;
  difficulty?: QuestionUpdateDifficulty;
  year?: number;
}