# lib/api-zod/src/generated/types/testInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * UPSC Prep Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TestInputStatus } from './testInputStatus';

export interface TestInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  /** @minLength 1 */
  subject: string;
  /** @minimum 1 */
  duration: number;
  /** @minimum 1 */
  totalMarks: number;
  /** @minimum 0 */
  negativeMarks?: number;
  status?: TestInputStatus;
  scheduledFor?: Date;
}