# lib/api-zod/src/generated/types/testUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * UPSC Prep Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TestUpdateStatus } from './testUpdateStatus';

export interface TestUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  subject?: string;
  /** @minimum 1 */
  duration?: number;
  /** @minimum 1 */
  totalMarks?: number;
  /** @minimum 0 */
  negativeMarks?: number;
  status?: TestUpdateStatus;
  scheduledFor?: Date;
}